Moved Store out of index
