+++ packages/ui/src/index.js
@@ -1,13 +1,11 @@
-import { render, createContext } from 'preact';
+import { render } from 'preact';
 import { html } from 'htm/preact';
 import { useState, useContext } from 'preact/hooks';
-import { autorun } from 'mobx';
-import { useLocalStore, useObserver } from 'mobx-react-lite';
-import { Router, Route, route } from 'preact-router';
+import { useObserver } from 'mobx-react-lite';
+import { Router, Route } from 'preact-router';
 import { createHashHistory } from 'history';
-import { JsonRpcMethod } from '@algosigner/common/messaging/types';
-
-import { sendMessage } from 'services/Messaging'
+import '@fortawesome/fontawesome-free/js/fontawesome'
+import '@fortawesome/fontawesome-free/js/solid'
 
 import Header from 'components/Header'
 import Footer from 'components/Footer'
@@ -23,62 +21,7 @@ import Account from 'pages/Account'
 import SendAlgos from 'pages/SendAlgos'
 import SignTransaction from 'pages/SignTransaction'
 
-import '@fortawesome/fontawesome-free/js/fontawesome'
-import '@fortawesome/fontawesome-free/js/solid'
-
-
-export const StoreContext = createContext();
-
-const StoreProvider = ({children}) => {
-  const existingStore = sessionStorage.getItem('wallet');
-  const store = useLocalStore(() => ({
-    ledger: 'MainNet',
-    setLedger: (ledger) => {
-      store.ledger = ledger;
-    },
-    updateWallet: (newWallet) => {
-      store.TestNet = newWallet.TestNet;
-      store.MainNet = newWallet.MainNet;
-    },
-    saveRequest: (request) => {
-      store.savedRequest = request;
-    },
-    clearSavedRequest: () => {
-      delete store.savedRequest;
-    },
-  }));
-
-  autorun(() => {
-    sessionStorage.setItem('wallet', JSON.stringify(store))
-  })
-
-  // Try to retrieve session from background
-  sendMessage(JsonRpcMethod.GetSession, {}, function(response) {
-    // Object.assign(store, JSON.parse(existingStore));
-    if (response && response.exist){
-      let hashPath = "";
-      if (window.location.hash.length > 0)
-         // Remove # from hash
-        hashPath = window.location.hash.slice(2);
-      if ('session' in response) {
-        store.updateWallet(response.session.wallet);
-        store.setLedger(response.session.ledger);
-        if (hashPath.length > 0) {
-          route(`/${hashPath}`)
-        } else {
-          route('/wallet')
-        }
-      } else {
-        route('/login/'+hashPath);
-      }
-    }
-  });
-
-  return html`
-    <${StoreContext.Provider} value=${store}>${children}</${StoreContext.Provider}>
-  `
-};
-
+import { StoreProvider } from 'services/StoreContext'
 
 require('./styles.scss');
