+++ packages/ui/src/pages/SendAlgos.ts
@@ -6,7 +6,7 @@ import { JsonRpcMethod } from '@algosigner/common/messaging/types';
 
 import { sendMessage } from 'services/Messaging'
 
-import { StoreContext } from 'index'
+import { StoreContext } from 'services/StoreContext'
 
 import HeaderView from 'components/HeaderView'
 import Authenticate from 'components/Authenticate'
