+++ packages/ui/src/pages/Wallet.ts
@@ -4,7 +4,7 @@ import { useState, useContext } from 'preact/hooks';
 import { useObserver } from 'mobx-react-lite';
 import { Link } from 'preact-router';
 
-import { StoreContext } from 'index'
+import { StoreContext } from 'services/StoreContext'
 import AccountPreview from 'components/AccountPreview'
